sql/parser: move LimitDecimalWidth to decimal.go where it belongs
